More problems

+++ deploy.sh
@@ -1,7 +1,11 @@
 #!/usr/bin/env bash
 echo 'Deployment in progress'
 pwd && cd target
-sudo systemctl stop grizzlystore-user.service || true
+
+if (( $(ps -ef | grep -v grep | grep grizzlystore-user | wc -l) > 0)) then
+	echo "User service is running...attempting to stop service!"
+	sudo systemctl stop grizzlystore-user.service || true
+fi
 
 if [ -f grizzlystore-user ] ; then
     sudo rm /etc/init.d/grizzlystore-user || true
